[intellij structure] refactoring: move LegacyPluginAnalysis to structure-intellij
This is needed to allow reusing this class in the Marketplace.
